refactor(Answer): 답변생성하기 및 답변수정하기 기능 리팩토링
